Code tidy up and new apk version

+++ app/src/main/java/com/example/taskapp/ui/TaskScreen.kt
@@ -1,8 +1,8 @@
 package com.example.taskapp.ui
 
-import androidx.compose.material3.*
-import androidx.compose.foundation.*
 import android.app.Application
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.background
 import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -20,11 +20,13 @@ import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SwipeToDismissBox
@@ -32,6 +34,7 @@ import androidx.compose.material3.SwipeToDismissBoxValue
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.rememberSwipeToDismissBoxState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -51,17 +54,16 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.taskapp.data.Category
-import com.example.taskapp.viewmodel.CategoriesVm
-import com.example.taskapp.viewmodel.UnifiedTaskViewModel
 import com.example.taskapp.data.Recurrence
 import com.example.taskapp.data.recurrenceLabel
+import com.example.taskapp.viewmodel.CategoriesVm
+import com.example.taskapp.viewmodel.UnifiedTaskViewModel
 import kotlinx.coroutines.launch
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
-import org.burnoutcrew.reorderable.rememberReorderableLazyListState
 import org.burnoutcrew.reorderable.ReorderableItem
-import org.burnoutcrew.reorderable.reorderable
 import org.burnoutcrew.reorderable.detectReorder
+import org.burnoutcrew.reorderable.rememberReorderableLazyListState
+import org.burnoutcrew.reorderable.reorderable
+import java.time.LocalDate
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
 @Composable
@@ -90,10 +92,10 @@ fun TasksScreen(
     val tasks by vm.tasksForList.collectAsState()
     val title by vm.title.collectAsState("")
     var showAdd by remember { mutableStateOf(false) }
-    val headerColor by vm.headerColor.collectAsState(0xFFEEEEEE.toLong())
+    val headerColor by vm.headerColor.collectAsState(0xFFEEEEEE)
     var askClear by remember { mutableStateOf(false) }
     val reorderState = rememberReorderableLazyListState(
-        onMove = { from, to -> vm.moveInList(tasks, from.index, to.index)}
+        onMove = { from, to -> vm.moveInList(tasks, from.index, to.index) }
 
 
     )
@@ -133,13 +135,17 @@ fun TasksScreen(
         Column(
             Modifier
                 .padding(pad)
-                .fillMaxSize()) {
-            val completedExists = tasks.any { it.completedDate != null && it.recurrence == Recurrence.NONE }
+                .fillMaxSize()
+        ) {
+            val completedExists =
+                tasks.any { it.completedDate != null && it.recurrence == Recurrence.NONE }
 
             Button(
                 onClick = { askClear = true },
                 enabled = completedExists,
-                modifier = Modifier.align(Alignment.End).padding(end = 16.dp, top = 8.dp)
+                modifier = Modifier
+                    .align(Alignment.End)
+                    .padding(end = 16.dp, top = 8.dp)
             ) { Text("Clear completed") }
 
             if (askClear) {
@@ -203,7 +209,8 @@ fun TasksScreen(
                             },
                             modifier = Modifier.padding(horizontal = 8.dp, vertical = 2.dp)
                         ) {
-                            val canDrag = dismissState.currentValue == SwipeToDismissBoxValue.Settled
+                            val canDrag =
+                                dismissState.currentValue == SwipeToDismissBoxValue.Settled
                             Card(
                                 modifier = Modifier
                                     .fillMaxWidth()
@@ -221,7 +228,8 @@ fun TasksScreen(
                                     verticalAlignment = Alignment.CenterVertically
                                 ) {
                                     val today = LocalDate.now().toString()
-                                    val doneToday = task.completedDate != null && task.completedDate >= today
+                                    val doneToday =
+                                        task.completedDate != null && task.completedDate >= today
 
                                     Text(
                                         text = task.text,
